Implement function get current processor ID
I want to implement a per-processor data structure. This requires a way
to get a unique identifier of the processor running the
currently-executing thread.

Implement a get_current_processor_id function for this task.

To test get_current_processor_id, correlate results with on-cpu and off-cpu
events captured using dtrace.