Refactor CDSChecker integration; remove workarounds
CDSChecker overrides <atomic>, <mutex>, and other headers. This has
caused many problems which needed tricky workarounds.

Instead of using CDSChecker's headers, create our own and call the C API
directly. This increases coupling, but I think the trade-off is worth
it.

CDSChecker does not have a C API for mutexes, so compile a wrapper which
uses CDSChecker's <mutex> (in a separate translation unit).