WIP Port cxxtrace to Linux+GCC
@nocommit test examples

Make cxxtrace's examples and automated tests compile and run on Linux
with GCC 9.1.0.

Because thread names are limited to 16 bytes on Linux, tweak the thread
names used in tests.

Future commits will implement some of the APIs, such as
processor_id_lookup, more efficiently. This commit just gets things up
and running.

This commit should not change behavior on macOS.