Avoid raw integer timestamps in test_processor_id_manual
Various parts of test_processor_id_manual assume that clock::sample is
integral. This is true now, but it might not be for other platforms
(e.g. Linux with clock_gettime). Refactor test_processor_id_manual so it
uses cxxtrace::time_point in most code.

This commit should not change behavior.