Group atomic, nonatomic, etc. into a class
Currently, the implementation of cxxtrace::detail::atomic,
cxxtrace::detail::nonatomic, and other templates is selected using a
type alias an #if:

    #if X
    template<class T> using atomic = x_atomic<T>;
    template<class T> using nonatomic = x_nonatomic<T>;
    #elif Y
    template<class T> using atomic = y_atomic<T>;
    template<class T> using nonatomic = y_nonatomic<T>;
    #endif

I want to instead select the implementation using a template parameter.
This is possible, but cumbersome:

    template<template<class> class Atomic, template<class> class Nonatomic>
    struct s
    {
      Atomic<int> a;
      Nonatomic<int> n;
    };
    using x_s = s<x_atomic, x_nonatomic>;

Simplify this pattern a bit by putting atomic-related classes and
functions into a parent class:

    template<class Synchronization>
    struct s
    {
      typename Synchronization::template atomic<int> a;
      typename Synchronization::template nonatomic<int> n;
    };
    using x_s = s<x_synchronization>;

(The typename and template keywords are necessary in C++17,
unfortunately, despite (to my knowledge) their absence causing no syntax
ambiguities.)

This commit should not change behavior.