Allow configuring processor count in rseq_scheduler
For some to-be-written tests, I want to constraint rseq_scheduler to
scheduling threads on only one processor. Allow tests to specify how
many processors should be reserved by adding a parameter to
rseq_scheduler' constructor.

With the current implementation, the processor count must be no less
than the number of threads. I plan to lift this restriction in a future
commit.