Replace rseq_scheduler array+size with vector
rseq_scheduler stores an array of processors and an integer indicating
how many processors are usable. This means rseq_scheduler manually
iterates over the array, and needs to be careful to not access unused
entries in the array. This is a bit cumbersome.

Replace rseq_scheduler's array and integer with an std::vector<>. This
simplies the code, but adds the minor cost of a few indirections.

This commit should not change behavior.