Fix test_current_processor_id_is_constant_within_critical_section
test_current_processor_id_is_constant_within_critical_section was
specified to run only on one processor. This makes it not test what it
was supposed to test. (If there's only one processor, of course the
processor ID will be constant!) Fix the test by running it on a
multiple-processor virtual machine.