Fix heap allocations in rseq_scheduler
Relacy runs many redundant iterations for tests using rseq_scheduler.
This is because rseq_scheduler's constructor allocates memory on the
heap.

Fix the redundant iterations by replacing the heap allocation with an
in-place allocation.

(I think this solution is ugly, but it's the simplest solution which
doesn't involve a new vector class or moving code into the header.)

The following table shows the improvement for
cxxtrace_test_rseq_scheduler_examples_relacy:

    |                                                   | Iterations     |
    | Test case                                         | Before | After |
    +---------------------------------------------------+--------+-------+
    | test_counter(processor_count=2)                   |   7132 |  3566 |
    | test_counter(processor_count=3)                   |  17544 |  8772 |
    | test_spin_lock<basic>(processor_count=2)          |   4108 |  2054 |
    | test_spin_lock<basic>(processor_count=3)          | 158136 | 79068 |
    | test_spin_lock<librseq_arm>(processor_count=2)    |   9980 |  4990 |
    | test_spin_lock<librseq_x86_64>(processor_count=2) |   4108 |  2054 |