Make using SCOPED_TRACE easier
SCOPED_TRACE is annoying to use because it accepts a string. Most uses
of SCOPED_TRACE are trying to log a value, so callers are forced to use
stringify explicitly.

Implement CXXTRACE_SCOPED_TRACE, which allows using iostreams for
messages (like gtest's assertion macros).

This commit should not change behavior.