Fix bad envp given to posix_spawnp
posix_spawnp accepts an environment ("envp"). If the environment pointer
is null, the behavior is unspecified. One some platforms, if envp is
null, the environment of the child becomes empty. An empty environment
is unintended.

When calling posix_spawnp, give an explicit environment pointer. In our
case, we want the child to inherit the full environment of the parent,
so give posix_spawnp environ (or *_NSGetEnviron() on macOS).