Create elf_chunk_range for iterating section data
Code which needs to scan the data for an ELF section needs to use the
elf_getdata API. This API is a bit verbose and is prone to errors such
as infinite loops.

Create elf_chunk_range and elf_chunk_iterator to simplify scanning an
ELF section. elf_chunk_range enables range-based for loops, for example.

This commit should not change behavior.